Submodules updated to version Onlyoffice v5.1

+++ core/ASCOfficeXlsFile2/source/XlsFormat/Logic/Biff_records/BopPop.cpp
@@ -1,5 +1,5 @@
 ﻿/*
- * (c) Copyright Ascensio System SIA 2010-2017
+ * (c) Copyright Ascensio System SIA 2010-2018
  *
  * This program is a free software product. You can redistribute it and/or
  * modify it under the terms of the GNU Affero General Public License (AGPL)
@@ -56,6 +56,7 @@ void BopPop::readFields(CFRecord& record)
 {
 	unsigned short flags;
 	record >> pst >> fAutoSplit >> split >> iSplitPos >> pcSplitPercent >> pcPie2Size >> pcGap >> numSplitValue >> flags;
+	
 	fHasShadow = GETBIT(flags, 0);
 }
 
@@ -102,7 +103,7 @@ int BopPop::serialize(std::wostream & _stream)
 		{
 			CP_XML_NODE(L"c:custSplit")
 			{
-				for (int i = 0 ; i < custom->rggrbit.pie_indices.size(); i++)
+				for (size_t i = 0 ; i < custom->rggrbit.pie_indices.size(); i++)
 				{
 					CP_XML_NODE(L"c:secondPiePt")
 					{
