+++ core/ASCOfficeXlsFile2/source/XlsFormat/Logic/Biff_records/MergeCells.cpp
@@ -1,5 +1,5 @@
 ﻿/*
- * (c) Copyright Ascensio System SIA 2010-2017
+ * (c) Copyright Ascensio System SIA 2010-2018
  *
  * This program is a free software product. You can redistribute it and/or
  * modify it under the terms of the GNU Affero General Public License (AGPL)
@@ -65,9 +65,8 @@ void MergeCells::readFields(CFRecord& record)
 int MergeCells::serialize(std::wostream & stream)
 {
 	CP_XML_WRITER(stream)    
-    {
-		
-        for (int i = 0 ; i < rgref.size(); i++)
+    {		
+        for (size_t i = 0 ; i < rgref.size(); i++)
 		{	
 			Ref8* ref = dynamic_cast<Ref8*>(rgref[i].get());
 			CP_XML_NODE(L"mergeCell")
