+++ core/ASCOfficeXlsFile2/source/XlsFormat/Logic/Biff_structures/ChartParsedFormula.cpp
@@ -1,5 +1,5 @@
 ﻿/*
- * (c) Copyright Ascensio System SIA 2010-2017
+ * (c) Copyright Ascensio System SIA 2010-2018
  *
  * This program is a free software product. You can redistribute it and/or
  * modify it under the terms of the GNU Affero General Public License (AGPL)
@@ -100,7 +100,7 @@ void ChartParsedFormula::load(CFRecord& record)
 
 bool ChartParsedFormula::inRange(const CellRef & ref)
 {
-	for (int i = 0 ; i < cell_ranges.size(); i++)
+	for (size_t i = 0 ; i < cell_ranges.size(); i++)
 	{
 		if (cell_ranges[i].inRange(ref))
 		{
